Integrate: [ 25754] Subject: [PATCH] Stas' croak patch and then some Message-ID: <20051013214212.GA1122@petdance.com> p4raw-link: @25754 on //depot/perl: 551405c409d33bc8cd0a20177c4ee21a204d18b5
p4raw-id: //depot/maint-5.8/perl@26686
p4raw-integrated: from //depot/perl@25754 'edit in' op.c (@25616..)
	pad.c (@25693..) perl.c (@25753..) 'merge in' regcomp.c
	(@25243..) gv.c (@25664..) toke.c (@25674..) embed.fnc proto.h
	(@25685..) utf8.c (@25716..) pp_sort.c (@25743..) sv.c
	(@25751..)

+++ regcomp.c
@@ -425,7 +425,7 @@ static const scan_data_t zero_scan_data =
 	MJD_OFFSET_DEBUG(("** (%d) offset of node %d is %d.\n",		\
 		__LINE__, (node), (byte)));				\
 	if((node) < 0) {						\
-	    Perl_croak(aTHX_ "value of node is %d in Offset macro", node); \
+	    Perl_croak(aTHX_ "value of node is %d in Offset macro", (int)(node)); \
 	} else {							\
 	    RExC_offsets[2*(node)-1] = (byte);				\
 	}								\
@@ -439,9 +439,9 @@ static const scan_data_t zero_scan_data =
 #define Set_Node_Length_To_R(node,len) STMT_START {			\
     if (! SIZE_ONLY) {							\
 	MJD_OFFSET_DEBUG(("** (%d) size of node %d is %d.\n",		\
-		__LINE__, (node), (len)));				\
+		__LINE__, (int)(node), (int)(len)));			\
 	if((node) < 0) {						\
-	    Perl_croak(aTHX_ "value of node is %d in Length macro", node); \
+	    Perl_croak(aTHX_ "value of node is %d in Length macro", (int)(node)); \
 	} else {							\
 	    RExC_offsets[2*(node)] = (len);				\
 	}								\
